CEDVA/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
gunicorn==20.1.0
psycopg2==2.9.5
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
unicodecsv==0.13.0
whitenoise==6.2.0
